Text Recognition (OCR)/OpenCV/requirements.txt
Python 3.0
OpenCV
PyTesseract using command pip install pytesseract
Pillow